fix: Improved the explanation of Wakers.
The original explanation of Wakers sort-of implied that there was
exactly one Waker for all futures (each executor creates one, then
passes clones of that to all registered futures).  The new phrasing
makes it clear that isn't the case.  Also added another couple of
paragraphs to the conclusion to really expand on that.